// src/Components/LibraryComponents/ChatComponent/CallPopUp/CallPopUp.types.ts
export interface IProps {
  injectComponent: any;
  patientId: string;
  registerEvent: any;
  unRegisterEvent: any;
  selectedClient: any;
  sessions: any;
  clientId: string;
  // userDetails?: Array<IUserDetails>;
  onClose: Function;
  tagOptions?: Array<ITag>;
  callScheduleMsg?: any;
  setCallScheduleMsg?: any;
  notifyEvent?: any;
  setIsAdd: Function;
  setMessagesList: Function;
}

export interface IUserDetails {
  userId: string;
  userName: string;
  roles: Array<string>;
  roleIds: Array<string>;
  isWatching?: boolean;
}
export interface ITag {
  id: string;
  label: string;
}
export interface IOptions {
  value: string;
  label: string;
}
export const TimeOptions = [
  '12:00 AM',
  '01:00 AM',
  '02:00 AM',
  '03:00 AM',
  '04:00 AM',
  '05:00 AM',
  '06:00 AM',
  '07:00 AM',
  '08:00 AM',
  '09:00 AM',
  '10:00 AM',
  '11:00 AM',
  '12:00 PM',
  '01:00 PM',
  '02:00 PM',
  '03:00 PM',
  '04:00 PM',
  '05:00 PM',
  '06:00 PM',
  '07:00 PM',
  '08:00 PM',
  '09:00 PM',
  '10:00 PM',
  '11:00 PM',
];